Update for negative values in loading from file

+++ Kalkulator_Macierzy/main.c
@@ -6,7 +6,7 @@
 struct Matrix {
 	int rows;
 	int columns;
-	int** data;
+	float** data;
 };
